Tienda-FullStack/src/main/java/com/tienda/tienda/controller/CategoriaController.java
package com.tienda.tienda.controller;
import com.tienda.tienda.repository.*;
import com.tienda.tienda.model.Categoria;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.*;
import java.util.List;





@RestController
@RequestMapping("/api/categorias")
@CrossOrigin(origins = "*")
public class CategoriaController {

    @Autowired
    private CategoriaRepository categoriaRepository;

    @GetMapping
    public List<Categoria> getAllCategorias() {
        return categoriaRepository.findByActivoTrue();
    }
}